requirements.txt
psycopg2-binary
pytz
python-decouple==3.1
django>=2.1.11
python-dotenv
django-multiselectfield
Pillow


